Do not parse HTTP headers in streaming mode
Header parsing code is now implemented for request_simple so we can act
as a simple HTTP <-> FCGI proxy. This means our streaming support is not
affected by the extra overhead of having to check for headers and buffer
enough data to parse them. If parsing headers is required in streaming
mode, it should be sufficient to identify the first returned chunk from
the body_reader iterator and parse chunk.stdout for HTTP headers (see
example code in _parse_headers()).

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>